server/requirements.txt
python-dotenv
flask
numpy
scipy
xmltodict
requests
shapely
